Reconfiguring EVERYTHING to use NewOptionBox
This commit makes ZERO visual changes (aside from the Effect titles, because I now know what they should be), but it's going to be a massive time save for me later.
Basically, it reconfigures everything so that instead of grammar data being stored in random arrays in App.vue, it's instead stored in grammardata.json.