Stress (Slot 10)
Added code that will add stress (and re-add cut vowels if necessary) to the word.
This means that the last thing that I need to add code that checks if slot 8 can be entirely dropped, and then I can call the word-generating algorithm complete! (Well, technically I also need to do the Slot 7/Slot 2 shortcuts...)
Also added code to allow the root to be inputted in uppercase or mixed case, if the user so desires.