Case modal tabs
Added tabs to the Case modal to switch between the various categories, rather than the awkward list of links at the top that it was originally.
Also fixed bugs regarding concatenation disabling relation (i.e. I forgot to code various checks in for that).
(Also replaced ALL instances of findIndex() with indexOf() because I didn't know that that existed, oops)