Better word-type selecting
Modified the word-type selecting so that instead of it being hidden in a not-very-well-explained dropdown menu in the settings, it's now right at the top of the page.
I did need to add the click-outside-v3 plugin to get the dropdowns working smoothly, and you have to click specifically on a dropdown button on mobile to close it, but other than that it's all fine.